backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==3.2.16
django-extra-fields==3.0.2
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
gunicorn==20.1.0
httpserver==1.1.0
idna==3.7
ingredients==0.3.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.9.0
pyshorteners==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
